+++ packages/ink-contracts/src/ink-client.ts
@@ -102,7 +102,7 @@ export interface InkClient<
   event: InkEventInterface<D["__types"]["event"]>
 }
 
-export const getInkClient = <
+const createInkClient = <
   D extends InkDescriptors<
     InkStorageDescriptor,
     InkCallableDescriptor,
@@ -163,6 +163,28 @@ export const getInkClient = <
         ? buildEventV4(builder.buildEvents)
         : buildEventV5(lookup, builder.buildEvent),
   }
+}
+
+const inkClientCache = new WeakMap<
+  InkDescriptors<any, any, any, any>,
+  InkClient<any>
+>()
+export const getInkClient = <
+  D extends InkDescriptors<
+    InkStorageDescriptor,
+    InkCallableDescriptor,
+    InkCallableDescriptor,
+    Event
+  >,
+>(
+  inkContract: D,
+): InkClient<D> => {
+  const cached = inkClientCache.get(inkContract)
+  if (cached) return cached
+
+  const client = createInkClient(inkContract)
+  inkClientCache.set(inkContract, client)
+  return client
 }
 
 const getAttributes = (spec: ConstructorSpec | MessageSpec) => ({